Initial commit
This is an experiment in using real-time raytracing to put the user
in a non-uniform non-Euclidean space.  The controls are WASD +
mouse look, but you can fly.

I think this once worked perfectly on Linux except that the sky was
initially black, but now there is a rendering error when you are in
the hypercylindrical part.